Create index.html
This is an HTML file that sets up a basic webpage for a React project. The webpage will display a quiz app, which will be rendered by the React application.

Here's a breakdown of the code:

- <!doctype html>: This is the document type declaration, which tells the browser that this is an HTML document.
- <html>, <head>, and <body>: These are the basic structural elements of an HTML document.
- <title>Quiz App</title>: This sets the title of the webpage, which appears in the browser's title bar and in search engine results.
- <div id="react-container"></div>: This is a container element where the React application will be rendered.
- <script src="app.js"></script>: This script tag links to the JavaScript file (app.js) that contains the React application code.

In summary, this HTML file sets up a basic webpage with a container element where the React application will be rendered, and links to the JavaScript file that contains the application code. The React application will display a quiz app in the container element.